+++ docs/source/changelog.rst
@@ -246,3 +246,8 @@ First started in 2009.
   :class:`cardinal_pythonlib.sqlalchemy.list_types.StringListType` about the
   slightly unusual behaviour of lists written to the database.
 - Moved to the ``create_all_autodocs.py`` system.
+
+**1.0.42**
+
+- Minor fix: ``__init__.py`` missing from ``cardinal_pythonlib/email``;
+  required for Python 3.5.